🏗️ Updated component directory, split out components
